[BUGFIX] Adds MM_opposite_field to allowed config entries for GroupProcessor
If someone configures a tca column as
            'type' => 'group',
            'internal_type' => 'db',
with a bidirectional MM relation, MM_opposite_field is not taken into account in Commonrepository->getLocalField() because the config MM_opposite_field is not allowed in GroupProcessor. Therefore this MM relation never gets updated if the editor is only changing on the opposite side of the relation.